% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{PreparescShaper}
\alias{PreparescShaper}
\alias{PreparescShaper.SingleCellExperiment}
\alias{PreparescShaper,SingleCellExperiment-method}
\title{Prepare \code{SingleCellExperiment} object for \code{scShaper} analysis}
\usage{
PreparescShaper.SingleCellExperiment(object, sparse.data)

\S4method{PreparescShaper}{SingleCellExperiment}(object, sparse.data = TRUE)
}
\arguments{
\item{object}{an object of \code{SingleCellExperiment} class}

\item{sparse.data}{a logical denoting if the data should
be converted into sparse format to save memory. Default is \code{TRUE}.}
}
\value{
an object of \code{SingleCellExperiment} class
}
\description{
This function prepares the \code{SingleCellExperiment} object for
\code{scShaper} analysis. The only required input is an object of class
\code{SingleCellExperiment} with data stored at least in the \code{logcounts} and \code{counts} slots.
}
\examples{
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays = list(counts=prosstt1$rna.raw,logcounts = prosstt1$rna.normalized))
sce <- PreparescShaper(sce)

}
\keyword{data}
\keyword{normalized}
\keyword{prepare}
\keyword{raw}
\keyword{scshaper}
