% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\docType{data}
\name{prosstt1}
\alias{prosstt1}
\title{Data simulated using PROSSTT simulator}
\format{
rna.normalized, rna.raw and pseudotime
}
\source{
\url{https://zenodo.org/record/1443566#.XiL7FyPRaUl}
}
\usage{
data(prosstt1)
}
\description{
Data simulated using PROSSTT simulator
}
\examples{
data(prosstt1)
}
\keyword{datasets}
