% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RunDimRed}
\alias{RunDimRed}
\alias{RunDimRed.SingleCellExperiment}
\alias{RunDimRed,SingleCellExperiment-method}
\title{Run dimensionality reduction.}
\usage{
RunDimRed.SingleCellExperiment(
  object,
  dim.red.method,
  num.pcs,
  num.tsne,
  tsne.perplexity
)

\S4method{RunDimRed}{SingleCellExperiment}(
  object,
  dim.red.method = "tsne",
  num.pcs = 50,
  num.tsne = 3,
  tsne.perplexity = 30
)
}
\arguments{
\item{object}{an object of \code{SingleCellExperiment} class}

\item{dim.red.method}{a character denoting which dimensionality reduction method to use.
Default is \code{tsne}. Other options are PCA (\code{pca}) 
and no dimensionality reduction (\code{none}).}

\item{num.pcs}{a positive integer denoting how many principal components (PCs) are extracted. 
Default is \code{50}.}

\item{num.tsne}{a positive integer denoting how many t-SNE dimensions to extract. 
Default is \code{3}.}

\item{tsne.perplexity}{a numeric denoting the perplexity of t-SNE. Default is \code{30}.}
}
\value{
an object of \code{SingleCellExperiment} class
}
\description{
This function performs dimensionality reduction using just PCA or PCA and t-SNE (default).
}
\examples{
library(SingleCellExperiment)
library(scShaper)
sce <- SingleCellExperiment(assays = list(counts=prosstt1$rna.raw,logcounts = prosstt1$rna.normalized))
sce <- PreparescShaper(sce)
sce <- RunDimRed(sce)
}
\keyword{PCA}
\keyword{Seurat}
\keyword{clustering}
\keyword{feature}
\keyword{graph}
\keyword{selection}
