% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RunKMeans}
\alias{RunKMeans}
\alias{RunKMeans.SingleCellExperiment}
\alias{RunKMeans,SingleCellExperiment-method}
\title{Run k-means clustering}
\usage{
RunKMeans.SingleCellExperiment(object, k.range, iter.max, nstart)

\S4method{RunKMeans}{SingleCellExperiment}(object, k.range = seq(10, 100), iter.max = 1000, nstart = 1)
}
\arguments{
\item{object}{an object of \code{SingleCellExperiment} class}

\item{k.range}{a numeric vector specifying the different cluster numbers, 
for which the discrete pseudotime is estimated using k-means and 
the special case of Kruskal's algorithm.
Default is \code{2:100}.}

\item{iter.max}{a positive integer denoting how iterations to perform 
at maximum in k-means. Default is \code{1000}.}

\item{nstart}{a positive integer denoting how many initializations to perform 
in k-means. Default is \code{1}. 
The third parameter to consider increasing if unsatisfactory results are obtained.}
}
\value{
an object of \code{SingleCellExperiment} class
}
\description{
This function runs k-means clustering for a range of k values.
}
\examples{
library(SingleCellExperiment)
library(scShaper)
sce <- SingleCellExperiment(assays = list(counts=prosstt1$rna.raw,logcounts = prosstt1$rna.normalized))
sce <- PreparescShaper(sce)
sce <- RunDimRed(sce)
sce <- RunKMeans(sce)
}
\keyword{PCA}
\keyword{Seurat}
\keyword{clustering}
\keyword{feature}
\keyword{graph}
\keyword{selection}
