% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RunLoessSmoothing}
\alias{RunLoessSmoothing}
\alias{RunLoessSmoothing.SingleCellExperiment}
\alias{RunLoessSmoothing,SingleCellExperiment-method}
\title{Run LOESS smoothing}
\usage{
RunLoessSmoothing.SingleCellExperiment(object, span)

\S4method{RunLoessSmoothing}{SingleCellExperiment}(object, span = 0.1)
}
\arguments{
\item{object}{an object of \code{SingleCellExperiment} class}

\item{span}{a numeric denoting the degree of smoothing in LOESS. 
Default is \code{0.1}. The second parameter to consider increasing 
if unsatisfactory results are obtained.}
}
\value{
an object of \code{SingleCellExperiment} class
}
\description{
This function performs LOESS to smoothen the crude initial average discrete pseudotime.
}
\examples{
library(SingleCellExperiment)
library(scShaper)
sce <- SingleCellExperiment(assays = list(counts=prosstt1$rna.raw,logcounts = prosstt1$rna.normalized))
sce <- PreparescShaper(sce)
sce <- RunDimRed(sce)
sce <- RunKMeans(sce)
sce <- RunSHP(sce)
sce <- RunPCAAnalysis(sce)
sce <- RunLoessSmoothing(sce)
}
\keyword{PCA}
\keyword{Seurat}
\keyword{clustering}
\keyword{feature}
\keyword{graph}
\keyword{selection}
