% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RunPCAAnalysis}
\alias{RunPCAAnalysis}
\alias{RunPCAAnalysis.SingleCellExperiment}
\alias{RunPCAAnalysis,SingleCellExperiment-method}
\title{Run PCA}
\usage{
RunPCAAnalysis.SingleCellExperiment(object, pca.rank)

\S4method{RunPCAAnalysis}{SingleCellExperiment}(object, pca.rank = 5)
}
\arguments{
\item{object}{an object of \code{SingleCellExperiment} class}

\item{pca.rank}{a positive integer denoting how many principal components to consider 
when generating the ensemble pseudotime. Default is \code{2}.}
}
\value{
an object of \code{SingleCellExperiment} class
}
\description{
This function performs PCA to analyze linear dependencies between the discrete pseudotimes.
}
\examples{
library(SingleCellExperiment)
library(scShaper)
sce <- SingleCellExperiment(assays = list(counts=prosstt1$rna.raw,logcounts = prosstt1$rna.normalized))
sce <- PreparescShaper(sce)
sce <- RunDimRed(sce)
sce <- RunKMeans(sce)
sce <- RunSHP(sce)
sce <- RunPCAAnalysis(sce)
}
\keyword{PCA}
\keyword{Seurat}
\keyword{clustering}
\keyword{feature}
\keyword{graph}
\keyword{selection}
