% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RunscShaper}
\alias{RunscShaper}
\alias{RunscShaper.SingleCellExperiment}
\alias{RunscShaper,SingleCellExperiment-method}
\title{Run \code{scShaper}.}
\usage{
RunscShaper.SingleCellExperiment(
  object,
  dim.red.method,
  num.pcs,
  num.tsne,
  tsne.perplexity,
  k.range,
  iter.max,
  nstart,
  pca.rank,
  span
)

\S4method{RunscShaper}{SingleCellExperiment}(
  object,
  dim.red.method = "tsne",
  num.pcs = 50,
  num.tsne = 3,
  tsne.perplexity = 30,
  k.range = seq(2, 100),
  iter.max = 1000,
  nstart = 1,
  pca.rank = 2,
  span = 0.1
)
}
\arguments{
\item{object}{an object of \code{SingleCellExperiment} class}

\item{dim.red.method}{a character denoting which dimensionality reduction method to use.
Default is \code{tsne}. Other options are PCA (\code{pca}) 
and no dimensionality reduction (\code{none}).}

\item{num.pcs}{a positive integer denoting how many principal components (PCs) are extracted. 
Default is \code{50}.}

\item{num.tsne}{a positive integer denoting how many t-SNE dimensions to extract. 
Default is \code{3}.}

\item{tsne.perplexity}{a numeric denoting the perplexity of t-SNE. Default is \code{30}.}

\item{k.range}{a numeric vector specifying the different cluster numbers, 
for which the discrete pseudotime is estimated using k-means and 
the special case of Kruskal's algorithm.
Default is \code{2:100}. 
The first parameter to consider increasing if unsatisfactory results are obtained.}

\item{iter.max}{a positive integer denoting how iterations to perform 
at maximum in k-means. Default is \code{1000}.}

\item{nstart}{a positive integer denoting how many initializations to perform 
in k-means. Default is \code{1}. 
The third parameter to consider increasing if unsatisfactory results are obtained.}

\item{pca.rank}{a positive integer denoting how many principal components to consider 
when generating the ensemble pseudotime. Default is \code{2}.}

\item{span}{a numeric denoting the degree of smoothing in LOESS. 
Default is \code{0.1}. The second parameter to consider increasing if unsatisfactory results are obtained.}
}
\value{
an object of \code{SingleCellExperiment} class
}
\description{
This function wraps all steps in scShaper from dimensionality reduction to LOESS smoothing.
}
\examples{
library(SingleCellExperiment)
library(scShaper)
sce <- SingleCellExperiment(assays = list(counts=prosstt1$rna.raw,logcounts = prosstt1$rna.normalized))
sce <- RunscShaper(sce)
}
\keyword{PCA}
\keyword{Seurat}
\keyword{clustering}
\keyword{feature}
\keyword{graph}
\keyword{selection}
