% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RunSHP}
\alias{RunSHP}
\alias{RunSHP.SingleCellExperiment}
\alias{RunSHP,SingleCellExperiment-method}
\title{Run a special case of kruskal's algorithm to find an optimal path through each clustering}
\usage{
RunSHP.SingleCellExperiment(object)

\S4method{RunSHP}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{an object of \code{SingleCellExperiment} class}
}
\value{
an object of \code{SingleCellExperiment} class
}
\description{
Run a special case of kruskal's algorithm to find an optimal path through each clustering:
shortest Hamiltonian path (SHP)
}
\examples{
library(SingleCellExperiment)
library(scShaper)
sce <- SingleCellExperiment(assays = list(counts=prosstt1$rna.raw,logcounts = prosstt1$rna.normalized))
sce <- PreparescShaper(sce)
sce <- RunDimRed(sce)
sce <- RunKMeans(sce)
sce <- RunSHP(sce)
}
\keyword{PCA}
\keyword{Seurat}
\keyword{clustering}
\keyword{feature}
\keyword{graph}
\keyword{selection}
